Create fragment from ItineraryCard
