+++ packages/components/index.js
@@ -27,7 +27,7 @@ export {
 export { PlaceCard } from './src/PlaceCard';
 export { BlackToAlpha };
 export { Shimmer } from './src/Shimmer';
-export { ItineraryCard, ItineraryCardProps } from './src/ItineraryCard';
+export { ItineraryCard, ItineraryCardType } from './src/ItineraryCard';
 export { TimelineFlightDetail } from './src/TimelineFlightDetail';
 export { default as Illustration } from './src/Illustration/Illustration';
 export {
