+++ packages/components/src/ItineraryCard/index.js
@@ -1,4 +1,6 @@
 // @flow
 
 export { default as ItineraryCard } from './ItineraryCard';
-export { ItineraryCardProps } from './ItineraryCardTypes';
+export type {
+  ItineraryCard as ItineraryCardType,
+} from './__generated__/ItineraryCard.graphql';
